Refactoring ready for merccator map
